+++ src/Symfony/Component/Mailer/Bridge/Amazon/Transport/SesApiTransport.php
@@ -25,19 +25,16 @@ use Symfony\Contracts\HttpClient\ResponseInterface;
  */
 class SesApiTransport extends AbstractApiTransport
 {
-    private const ENDPOINT = 'https://email.%region%.amazonaws.com';
-
-    private $accessKey;
-    private $secretKey;
+    private $credential;
     private $region;
 
     /**
-     * @param string $region Amazon SES region (currently one of us-east-1, us-west-2, or eu-west-1)
+     * @param ApiTokenCredential|UsernamePasswordCredential $credential credential object for SES authentication. ApiTokenCredential and UsernamePasswordCredential are supported.
+     * @param string                                        $region     Amazon SES region (currently one of us-east-1, us-west-2, or eu-west-1)
      */
-    public function __construct(string $accessKey, string $secretKey, string $region = null, HttpClientInterface $client = null, EventDispatcherInterface $dispatcher = null, LoggerInterface $logger = null)
+    public function __construct($credential, string $region = null, HttpClientInterface $client = null, EventDispatcherInterface $dispatcher = null, LoggerInterface $logger = null)
     {
-        $this->accessKey = $accessKey;
-        $this->secretKey = $secretKey;
+        $this->credential = $credential;
         $this->region = $region ?: 'eu-west-1';
 
         parent::__construct($client, $dispatcher, $logger);
@@ -45,23 +42,27 @@ class SesApiTransport extends AbstractApiTransport
 
     public function getName(): string
     {
-        return sprintf('api://%s@ses?region=%s', $this->accessKey, $this->region);
+        $login = null;
+        if ($this->credential instanceof ApiTokenCredential) {
+            $login = $this->credential->getAccessKey();
+        } else {
+            $login = $this->credential->getUsername();
+        }
+
+        return sprintf('api://%s@ses?region=%s', $login, $this->region);
     }
 
     protected function doSendApi(Email $email, SmtpEnvelope $envelope): ResponseInterface
     {
-        $date = gmdate('D, d M Y H:i:s e');
-        $auth = sprintf('AWS3-HTTPS AWSAccessKeyId=%s,Algorithm=HmacSHA256,Signature=%s', $this->accessKey, $this->getSignature($date));
+        $request = new SesRequest($this->client, $this->region);
+        $request->setMode(SesRequest::REQUEST_MODE_API);
+        $request->setCredential($this->credential);
 
-        $endpoint = str_replace('%region%', $this->region, self::ENDPOINT);
-        $response = $this->client->request('POST', $endpoint, [
-            'headers' => [
-                'X-Amzn-Authorization' => $auth,
-                'Date' => $date,
-                'Content-Type' => 'application/x-www-form-urlencoded',
-            ],
-            'body' => $this->getPayload($email, $envelope),
-        ]);
+        if ($email->getAttachments()) {
+            $response = $request->sendRawEmail($email->toString());
+        } else {
+            $response = $request->sendEmail($email, $envelope);
+        }
 
         if (200 !== $response->getStatusCode()) {
             $error = new \SimpleXMLElement($response->getContent(false));
@@ -70,42 +71,5 @@ class SesApiTransport extends AbstractApiTransport
         }
 
         return $response;
-    }
-
-    private function getSignature(string $string): string
-    {
-        return base64_encode(hash_hmac('sha256', $string, $this->secretKey, true));
-    }
-
-    private function getPayload(Email $email, SmtpEnvelope $envelope): array
-    {
-        if ($email->getAttachments()) {
-            return [
-                'Action' => 'SendRawEmail',
-                'RawMessage.Data' => base64_encode($email->toString()),
-            ];
-        }
-
-        $payload = [
-            'Action' => 'SendEmail',
-            'Destination.ToAddresses.member' => $this->stringifyAddresses($this->getRecipients($email, $envelope)),
-            'Message.Subject.Data' => $email->getSubject(),
-            'Source' => $envelope->getSender()->toString(),
-        ];
-
-        if ($emails = $email->getCc()) {
-            $payload['Destination.CcAddresses.member'] = $this->stringifyAddresses($emails);
-        }
-        if ($emails = $email->getBcc()) {
-            $payload['Destination.BccAddresses.member'] = $this->stringifyAddresses($emails);
-        }
-        if ($email->getTextBody()) {
-            $payload['Message.Body.Text.Data'] = $email->getTextBody();
-        }
-        if ($email->getHtmlBody()) {
-            $payload['Message.Body.Html.Data'] = $email->getHtmlBody();
-        }
-
-        return $payload;
     }
 }