+++ src/Symfony/Component/Mailer/Bridge/Amazon/Transport/SesHttpTransport.php
@@ -12,6 +12,9 @@
 namespace Symfony\Component\Mailer\Bridge\Amazon\Transport;
 
 use Psr\Log\LoggerInterface;
+use Symfony\Component\Mailer\Bridge\Amazon\Credential\ApiTokenCredential;
+use Symfony\Component\Mailer\Bridge\Amazon\Credential\UsernamePasswordCredential;
+use Symfony\Component\Mailer\Bridge\Amazon\SesRequest;
 use Symfony\Component\Mailer\Exception\HttpTransportException;
 use Symfony\Component\Mailer\SentMessage;
 use Symfony\Component\Mailer\Transport\AbstractHttpTransport;
@@ -24,19 +27,16 @@ use Symfony\Contracts\HttpClient\ResponseInterface;
  */
 class SesHttpTransport extends AbstractHttpTransport
 {
-    private const ENDPOINT = 'https://email.%region%.amazonaws.com';
-
-    private $accessKey;
-    private $secretKey;
+    private $credential;
     private $region;
 
     /**
-     * @param string $region Amazon SES region (currently one of us-east-1, us-west-2, or eu-west-1)
+     * @param ApiTokenCredential|UsernamePasswordCredential $credential credential object for SES authentication. ApiTokenCredential and UsernamePasswordCredential are supported.
+     * @param string                                        $region     Amazon SES region (currently one of us-east-1, us-west-2, or eu-west-1)
      */
-    public function __construct(string $accessKey, string $secretKey, string $region = null, HttpClientInterface $client = null, EventDispatcherInterface $dispatcher = null, LoggerInterface $logger = null)
+    public function __construct($credential, string $region = null, HttpClientInterface $client = null, EventDispatcherInterface $dispatcher = null, LoggerInterface $logger = null)
     {
-        $this->accessKey = $accessKey;
-        $this->secretKey = $secretKey;
+        $this->credential = $credential;
         $this->region = $region ?: 'eu-west-1';
 
         parent::__construct($client, $dispatcher, $logger);
@@ -44,25 +44,23 @@ class SesHttpTransport extends AbstractHttpTransport
 
     public function getName(): string
     {
-        return sprintf('http://%s@ses?region=%s', $this->accessKey, $this->region);
+        $login = null;
+        if ($this->credential instanceof ApiTokenCredential) {
+            $login = $this->credential->getAccessKey();
+        } else {
+            $login = $this->credential->getUsername();
+        }
+
+        return sprintf('http://%s@ses?region=%s', $login, $this->region);
     }
 
     protected function doSendHttp(SentMessage $message): ResponseInterface
     {
-        $date = gmdate('D, d M Y H:i:s e');
-        $auth = sprintf('AWS3-HTTPS AWSAccessKeyId=%s,Algorithm=HmacSHA256,Signature=%s', $this->accessKey, $this->getSignature($date));
+        $request = new SesRequest($this->client, $this->region);
+        $request->setMode(SesRequest::REQUEST_MODE_HTTP);
+        $request->setCredential($this->credential);
 
-        $endpoint = str_replace('%region%', $this->region, self::ENDPOINT);
-        $response = $this->client->request('POST', $endpoint, [
-            'headers' => [
-                'X-Amzn-Authorization' => $auth,
-                'Date' => $date,
-            ],
-            'body' => [
-                'Action' => 'SendRawEmail',
-                'RawMessage.Data' => base64_encode($message->toString()),
-            ],
-        ]);
+        $response = $request->sendRawEmail($message->toString());
 
         if (200 !== $response->getStatusCode()) {
             $error = new \SimpleXMLElement($response->getContent(false));
@@ -71,10 +69,5 @@ class SesHttpTransport extends AbstractHttpTransport
         }
 
         return $response;
-    }
-
-    private function getSignature(string $string): string
-    {
-        return base64_encode(hash_hmac('sha256', $string, $this->secretKey, true));
     }
 }