+++ src/Symfony/Component/Mailer/Bridge/Amazon/Transport/SesSmtpTransport.php
@@ -12,6 +12,7 @@
 namespace Symfony\Component\Mailer\Bridge\Amazon\Transport;
 
 use Psr\Log\LoggerInterface;
+use Symfony\Component\Mailer\Bridge\Amazon\Credential\UsernamePasswordCredential;
 use Symfony\Component\Mailer\Transport\Smtp\EsmtpTransport;
 use Symfony\Contracts\EventDispatcher\EventDispatcherInterface;
 
@@ -21,13 +22,14 @@ use Symfony\Contracts\EventDispatcher\EventDispatcherInterface;
 class SesSmtpTransport extends EsmtpTransport
 {
     /**
-     * @param string $region Amazon SES region (currently one of us-east-1, us-west-2, or eu-west-1)
+     * @param UsernamePasswordCredential $credential credential object for SES authentication
+     * @param string                     $region     Amazon SES region (currently one of us-east-1, us-west-2, or eu-west-1)
      */
-    public function __construct(string $username, string $password, string $region = null, EventDispatcherInterface $dispatcher = null, LoggerInterface $logger = null)
+    public function __construct(UsernamePasswordCredential $credential, string $region = null, EventDispatcherInterface $dispatcher = null, LoggerInterface $logger = null)
     {
         parent::__construct(sprintf('email-smtp.%s.amazonaws.com', $region ?: 'eu-west-1'), 587, true, $dispatcher, $logger);
 
-        $this->setUsername($username);
-        $this->setPassword($password);
+        $this->setUsername($credential->getUsername());
+        $this->setPassword($credential->getPassword());
     }
 }